fix: 사이드바 렌더링 관련 수정
